requirements.txt
albumentations==1.3.0
gradio==3.32.0
keras==2.12.0
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
